Separated Java and Scala APIs
